requirements.txt
# Core runtime
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-CORS==4.0.0
flask-restx==1.2.0
python-dotenv==1.0.0
requests==2.31.0

# Frontend (Streamlit UI)
streamlit==1.28.1
pandas==2.1.3
plotly==5.17.0

# Background tasks (optional)
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
